requirements.txt
azure-storage-file-share
requests
